requirements.txt
about-time==3.1.1
alive-progress==2.1.0
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1635819660018/work
argcomplete @ file:///home/conda/feedstock_root/build_artifacts/argcomplete_1619128689661/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
Bottleneck==1.3.2
certifi==2021.10.8
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1636043402768/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
grapheme==0.6.0
greenlet==1.1.2
IMDbPY==2021.4.18
importlib-metadata @ file:///Users/runner/miniforge3/conda-bld/importlib-metadata_1636431629957/work
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1636998649625/work/dist/ipykernel-6.5.0-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1637978887333/work
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1637175421629/work
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1636814373394/work
lxml==4.6.4
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mdutils==1.3.1
mkl-fft==1.3.1
mkl-random @ file:///opt/concourse/worker/volumes/live/133f1d0b-8fd8-4fa0-679e-7214401cbd75/volume/mkl_random_1626186083386/work
mkl-service==2.4.0
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
numexpr @ file:///opt/concourse/worker/volumes/live/6bbc9b0e-ade0-40db-6ee6-74c979385fbb/volume/numexpr_1618856531941/work
numpy @ file:///opt/concourse/worker/volumes/live/046e897c-3622-47d5-79cf-02f72d341f0d/volume/numpy_and_numpy_base_1634106712173/work
pandas==1.3.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1636045889479/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1635877593895/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy==1.4.27
tabulate==0.8.9
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1635819700605/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1637155965157/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1633302054558/work
